Refactor voice interface to use state lifting pattern
Resolves voice interface issues with ref-based state management and
component remounting inconsistencies by implementing Option 1 approach:
State Lifting + Custom Hook pattern.

Changes:
- Created useSpeechRecognition custom hook to encapsulate speech logic
- Lifted speech recognition state from VoiceInterface to App component
- Simplified VoiceInterface to be purely presentational (740→140 lines)
- Removed all ref-based state management (recognitionRef, synthRef, etc.)
- Eliminated debugging code blocks (if false && onListeningChange)
- Added handleVoiceTranscript function at App level for processing
- Integrated with existing Anthropic streaming hook
- Speech recognition now persists across component remounts
- Consistent parent-child state synchronization

Benefits:
- No more lost speech sessions during component remounts
- Single source of truth for speech recognition state
- Better separation of concerns between UI and logic
- Improved robustness for production use
- Eliminated debugging flags and inconsistent callbacks
